% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readSMI}
\alias{readSMI}
\title{readSMI}
\usage{
readSMI(file)
}
\arguments{
\item{file}{smi filename to read.}
}
\description{
This function reads SMI eye-tracker files
}
\examples{
data = readSMI("AXCPT_Subject_1 Samples.txt")
}

