% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSMI.R
\name{readSMI}
\alias{readSMI}
\title{readSMI}
\usage{
readSMI(file, msgcolposition = 4, msgprefix = "# Message: ",
  eventsfile = NULL)
}
\arguments{
\item{file}{smi filename to read.}

\item{msgcolposition}{column of MSG data (optional)}

\item{msgprefix}{prefix to remove from MSG data (optional)}

\item{eventsfile}{events file for merging with main file (optional / experimental)}
}
\description{
This function reads SMI eye-tracker files
}
\examples{
data = readSMI("AXCPT_Subject_1 Samples.txt")
}

